+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/service/ManagedRelationUpsertService.kt
@@ -20,7 +20,7 @@
 package org.eclipse.tractusx.bpdm.pool.service
 
 import org.eclipse.tractusx.bpdm.pool.api.model.DataSpaceParticipantDto
-import org.eclipse.tractusx.bpdm.pool.api.model.RelationType
+import org.eclipse.tractusx.bpdm.pool.api.model.LegalEntityRelationType
 import org.eclipse.tractusx.bpdm.pool.api.model.request.DataSpaceParticipantUpdateRequest
 import org.eclipse.tractusx.bpdm.pool.dto.UpsertResult
 import org.eclipse.tractusx.bpdm.pool.entity.LegalEntityDb
@@ -45,7 +45,7 @@ class ManagedRelationUpsertService(
         validateNoChain(upsertRequest)
 
         val result = relationUpsertService.upsertRelation(
-            RelationUpsertService.UpsertRequest(upsertRequest.source, upsertRequest.target, RelationType.IsManagedBy, upsertRequest.validityPeriods)
+            RelationUpsertService.UpsertRequest(upsertRequest.source, upsertRequest.target, LegalEntityRelationType.IsManagedBy, upsertRequest.validityPeriods)
         )
         makeManagedEntityParticipantIfRequired(upsertRequest.source)
 
@@ -55,7 +55,7 @@ class ManagedRelationUpsertService(
     private fun validateSingleManager(upsertRequest: IRelationUpsertStrategyService.UpsertRequest){
         val source = upsertRequest.source
 
-        val existingManagers = relationRepository.findByTypeAndStartNode(RelationType.IsManagedBy, source)
+        val existingManagers = relationRepository.findByTypeAndStartNode(LegalEntityRelationType.IsManagedBy, source)
         val overlappingExistingManagers = relationUpsertService.filterOverlappingRelations(upsertRequest, existingManagers)
 
         if (overlappingExistingManagers.isNotEmpty()) {
@@ -71,8 +71,8 @@ class ManagedRelationUpsertService(
         val source = upsertRequest.source
         val target = upsertRequest.target
 
-        val sourceRelations = relationRepository.findInSourceOrTarget(RelationType.IsManagedBy, source)
-        val targetRelations = relationRepository.findInSourceOrTarget(RelationType.IsManagedBy, target)
+        val sourceRelations = relationRepository.findInSourceOrTarget(LegalEntityRelationType.IsManagedBy, source)
+        val targetRelations = relationRepository.findInSourceOrTarget(LegalEntityRelationType.IsManagedBy, target)
 
         val overlappingSourceRelations = relationUpsertService.filterOverlappingRelations(upsertRequest, sourceRelations)
         val overlappingTargetRelations = relationUpsertService.filterOverlappingRelations(upsertRequest, targetRelations)